CPM-935: Merge migrations
