+++ upgrades/schema/Version_8_0_20230110131126_update_target_column_for_identifier_generators.php
@@ -35,6 +35,10 @@ ALTER TABLE pim_catalog_identifier_generator ADD COLUMN target_id INT NOT NULL A
 UPDATE pim_catalog_identifier_generator SET target_id=(SELECT id FROM pim_catalog_attribute WHERE code=target);
 ALTER TABLE pim_catalog_identifier_generator ADD CONSTRAINT `pim_catalog_identifier_generator_ibfk_1` FOREIGN KEY(target_id) REFERENCES pim_catalog_attribute(id) ON DELETE CASCADE;
 ALTER TABLE pim_catalog_identifier_generator DROP COLUMN target;
+
+ALTER TABLE pim_catalog_identifier_generator ADD COLUMN options JSON NOT NULL AFTER delimiter;
+UPDATE pim_catalog_identifier_generator SET options=JSON_OBJECT('delimiter', delimiter, 'text_transformation', 'no');
+ALTER TABLE pim_catalog_identifier_generator DROP COLUMN delimiter;
 SQL;
 
         $this->addSql($sql);
